Fixed issues with adding users to groups

+++ app/Http/Controllers/GroupController.php
@@ -40,7 +40,11 @@ class GroupController extends Controller
         $user = User::where('email', $userString)->first() ?? User::where('name', $userString)->first();
     
         if (!$user) {
-            return response()->json(['error' => 'User not found'], 404);
+            return response()->json(['error' => 'User not found']);
+        }
+
+        if ($group->users->contains($user)) {
+            return response()->json(['error' => 'User already in group']);
         }
 
         $group->addUser($user);
@@ -54,7 +58,11 @@ class GroupController extends Controller
         $user = User::find($userId);
 
         if (!$user) {
-            return response()->json(['error' => 'User not found'], 404);
+            return response()->json(['error' => 'User not found']);
+        }
+
+        if (! $group->users->contains($user)) {
+            return response()->json(['error' => 'User not in group']);
         }
 
         $group->users()->detach($user);
